Add basic HTTP client
This is useful to download pretrained tokenizers and remove the need to
use the feature from the native code. With that we could also remove
the "OpenSSL" dependency that is needed for the "http" feature of
the "tokenizers" crate.

Most of the request details is based on the original implementation at
https://github.com/huggingface/tokenizers/blob/102dfe87a3bf590dcb888f4f4644ad32def8c464/tokenizers/src/utils/from_pretrained.rs

The HTTP client is inspired by Req, and in fact can be replaced by Req
if desired.