Add riscv64gc-unknown-linux-gnu for RISC-V Nerves devices (#27)
Nerves supports 64-bit RISC-V devices. One of these is the MangoPi MQ
Pro which is being used as an alternative to the Raspberry Pi. Currently
it's much, much easier to use precompiled Rust binaries with Nerves. In
fact, manually crosscompiling `tokenizers` with Nerves doesn't work due
to a compiler error deep in the Rust build.

This adds the riscv64gc target to make it possible to use `tokenizers`
with Nerves on 64-bit RISC-V devices as easy as 64-bit ARM ones. This,
in turn, enables Bumblebee support.